Fix SD config loading lifetime and SdFat file handling
- load SD YAML into runtime config vars instead of assigning array globals
- validate required YAML keys before applying overrides
- use SdFat FsFile/O_RDONLY for config reads
- align writeFile signature with header to fix link mismatch

+++ src/file_utils.cpp
@@ -20,7 +20,7 @@ extern Inkplate board;
   - ESP_ERR_EFILEW if number of retries is exceeded without success.
 */
 
-esp_err_t writeFile(uint8_t* buf, int32_t size, const char* filePath) {
+esp_err_t writeFile(uint8_t* buf, size_t size, const char* filePath) {
     logf(LOG_DEBUG, "writing file to path %s", filePath);
     SdFat &sd = board.getSdFat();
